Version 0.4.0
The query module can be used for interacting with feeds to retrieve video objects.

The Dispatcher class got a new method, some debug statements were cleaned up, more tests were written, and most importantly the README has been updated to have some API documentation in it. Long way to go, but I'm going to make it a requirement when submitting a new module.

Signed-off-by: Brandon Aaskov <aaskov@gmail.com>