Updated to use an array of player objects
Now we pass around player objects, and we store all of those in
_players, which is an array (again). This is cleaner. I like this. It's
easier to inspect player objects and much clearer as to what's getting
passed around. We also allow for the selector that gets the controller
to be either the iframe id or the id that was supplied (either by
javascript injection or attribute declaration).