Refactored iframe creation
Previously, it was way too complicated - the mere fact that I had
cyclical dependencies in iframe.js was evident of that. Now Iframe is a
contstructor that we can create new instances of. This helps us manage
the players better since everything is controlled in player.js, and
everything lives in the _players array. We're also now following the
event flow more precisely so that we don't call bind on the pdk
before/after we should.